Update formSubmitTransformers to use v2 logic and remove v1 implementation
